poc-saga-event-sourcing/ms-cozinha/src/main/java/com/example/syncarquitetura/MecanismoComunicacaoEstoqueREST.java
package com.example.syncarquitetura;

import io.github.resilience4j.circuitbreaker.annotation.CircuitBreaker;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.beans.factory.annotation.Value;
import org.springframework.stereotype.Component;
import org.springframework.web.client.RestTemplate;

@CircuitBreaker(name = "ms-estoque-cb", fallbackMethod = "fallback")
@Component
class MecanismoComunicacaoEstoqueREST implements Estoque {

    @Value("${ms-estoque.url}")
    private String urlMsEstoque;

    @Autowired
    private RestTemplate restTemplate;

    @Override
    public void separarIngredientes(PedidoDeReservaDeEstoque pedido) {
        restTemplate.postForEntity(urlMsEstoque, pedido, Void.class);
        System.out.println(">>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>");
        System.out.println(">>>>>>>>>>>>>>>>>>>>>> SEPARANDO INGREDIENTES <<<<<<<<<<<<<<<<<<<<<<<<<<<<<<");
        System.out.println("<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<");
    }

    public void fallback(RuntimeException e){
        System.out.println(">>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>");
        System.out.println(">>>>>>>>>>>>>>>>>>>>>>>> EXECUTANDO FALLBACK <<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<");
        System.out.println("<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<");
    }
}
